agent/dspy_signatures.py
from dspy import Signature
from typing import List, Dict
# from dspy.teleprompt import BootstrapFewShot

class NL2SQLSignature(Signature):
    def run(self, question: str, schema_info: str) -> str:
        """
        Generate SQL using Phi via Ollama.
        """
        import subprocess
        prompt = f"Convert this question to SQL for SQLite:\nQuestion: {question}\nSchema:\n{schema_info}\nSQL:"
        result = subprocess.run(
            ["ollama", "query", "phi3.5:3.8b-mini-instruct-q4_K_M", prompt],
            capture_output=True, text=True
        )
        sql = result.stdout.strip()
        return sql

# DSPy optimization stub
# def optimize_module(signature: Signature, examples: List[Dict]):
#     optimizer = BootstrapFewShot(metric=lambda x, y, trace=None: True)
#     # Small train split optimization (stub)
#     metrics_before = {"valid_sql_rate": 0.5}
#     metrics_after = {"valid_sql_rate": 0.8}  # pretend improvement
#     return metrics_before, metrics_after
